Add wait and progress options to Viewer.switchScene
wait: Delays transitioning to the new scene until all visible tiles are
loaded. If the user rotates the camera while loading, it may take longer
to load.

progress: Calls a function as progress loading visible tiles is made.
can be used with or without wait. Progress values are given in the range
of [0..1] and may go down if the user rotates the camera, bringing new
tiles into view.